+++ pysatMadrigal/instruments/__init__.py
@@ -6,6 +6,7 @@
 """Import the Instrument sub-modules and methods."""
 # Import Madrigal instruments
 from pysatMadrigal.instruments import dmsp_ivm
+from pysatMadrigal.instruments import dmsp_ssj
 from pysatMadrigal.instruments import gnss_tec
 from pysatMadrigal.instruments import jro_isr
 from pysatMadrigal.instruments import madrigal_pandas
@@ -14,4 +15,4 @@ from pysatMadrigal.instruments import madrigal_pandas
 from pysatMadrigal.instruments import methods  # noqa F401
 
 # Define variable name with all available instruments
-__all__ = ['dmsp_ivm', 'gnss_tec', 'jro_isr', 'madrigal_pandas']
+__all__ = ['dmsp_ivm', 'dmsp_ssj', 'gnss_tec', 'jro_isr', 'madrigal_pandas']